Move response file handling to separate module
* Moves the MSVC response file code into a new module
* Makes no† changes to the response file code

† Had to make `SplitMsvcResponseFileArgs` `pub`, so that it can be used.
