+++ src/compiler/mod.rs
@@ -23,6 +23,7 @@ mod diab;
 mod gcc;
 mod msvc;
 mod nvcc;
+mod response_file;
 mod rust;
 mod tasking_vx;
 #[macro_use]
